#4: Fixed few issues with refactoring
